# [Perf improvement] Fixing Content.Medium creation for the unwrapping of custom modifiers.