@bachand 2023-12-25
Thank you for building such a useful library, @nalexn ! We are happy to contribute.